requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiolimiter==1.2.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
backoff==2.2.1
bce-python-sdk==0.9.25
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.15
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
durationpy==0.9
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
ffmpy==0.5.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google_search_results==2.4.2
googleapis-common-protos==1.66.0
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.1.1
grpcio==1.68.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.3.13
langchain-chroma==0.1.4
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.2.11
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
narwhals==1.20.1
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.58.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
ormsgpack==1.9.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pillow==11.1.0
posthog==3.7.4
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rsa==4.9
ruff==0.8.5
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
